+++ src/feat-mypage/components/tabs/index.ts
@@ -3,5 +3,5 @@ export { default as ReservationsTab } from './ReservationsTab';
 export { default as CouponsTab } from './CouponsTab';
 export { default as SettlementsTab } from './SettlementsTab';
 export { default as PointHistoryTab } from './PointHistoryTab';
-export { default as PerformanceManagementTab } from './PerformanceManagementTab';
+export { default as PerformanceHostDashboard } from './PerformanceHostDashboard';
 export { default as ScrapPage } from './ScrapPage';
