+++ src/feat-mypage/styles/PerformanceHostPage.css
@@ -188,10 +188,10 @@
 }
 
 /* 공연 이미지 */
-.performance-image {
+.performance-create-image {
   position: relative;
-  height: 300px !important;
-  width: 290px !important;
+  width: 100%;
+  height: 280px;
   overflow: hidden;
   display: flex;
   align-items: center;
@@ -199,11 +199,11 @@
   text-align: center;
 }
 
-.performance-img {
+.performance-create-img { 
   width: 100%;
-  height: 100%;
+  height: 280px;
   object-fit: cover;
-  border-radius: 8px 8px 0 0;
+  border-radius: 12px 12px 0 0;
 }
 
 .no-image {
