+++ src/home/pages/Performance/PerformanceHostPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Layout from '../../../components/layout/Layout';
-import PerformanceHostDashboard from '../../components/performance/PerformanceHostDashboard';
+import PerformanceHostDashboard from '../../../feat-mypage/components/tabs/PerformanceHostDashboard';
 
 const PerformanceHostPage: React.FC = () => {
   return (
